# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bakery ERP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /overtime-rates:
    get:
      operationId: listOvertimeRates
      tags: [overtime]
      summary: List overtime hourly rate history
      parameters:
        - name: branchId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OvertimeRate"
    post:
      operationId: createOvertimeRate
      tags: [overtime]
      summary: Set a new overtime hourly rate effective from a date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OvertimeRateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OvertimeRate"

  # ---------- BONUSES ----------
...
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List loans / advances
      parameters:
        - name: employeeId
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Add a loan / advance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}:
    patch:
      operationId: updateLoan
      tags: [loans]
      summary: Update a loan (e.g. record installment payment)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Not found
    delete:
      operationId: deleteLoan
      tags: [loans]
      summary: Delete a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ---------- PAYROLL ----------
  /payroll:
    get:
      operationId: listPayroll
      tags: [payroll]
      summary: List generated payroll records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payroll"
  /payroll/generate:
    post:
      operationId: generatePayroll
      tags: [payroll]
      summary: Generate/finalize a payroll record for an employee for a given month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollGenerateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
  /payroll/current/{employeeId}:
    get:
      operationId: getCurrentPayroll
      tags: [payroll]
      summary: Get the live, accruing payroll estimate for the current month
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentPayroll"
        "404":
          description: Not found

  # ---------- WAREHOUSE ----------
...
components:
  schemas:
...
    OvertimeRate:
      type: object
      required: [id, branchId, ratePerHour, effectiveFrom]
      properties:
        id: { type: string }
        branchId: { type: string }
        ratePerHour: { type: number }
        effectiveFrom: { type: string }
    OvertimeRateInput:
      type: object
      required: [branchId, ratePerHour, effectiveFrom]
      properties:
        branchId: { type: string }
        ratePerHour: { type: number, minimum: 0 }
        effectiveFrom: { type: string }
...
    Loan:
      type: object
      required: [id, employeeId, amount, date, installment, remaining]
      properties:
        id: { type: string }
        employeeId: { type: string }
        employeeName: { type: ["string", "null"] }
        amount: { type: number }
        date: { type: string }
        installment: { type: number }
        remaining: { type: number }
    LoanInput:
      type: object
      required: [employeeId, amount, date, installment]
      properties:
        employeeId: { type: string }
        amount: { type: number, minimum: 0 }
        date: { type: string }
        installment: { type: number, minimum: 0 }
    LoanUpdate:
      type: object
      properties:
        installment: { type: number, minimum: 0 }
        remaining: { type: number, minimum: 0 }
    PayrollStatus:
      type: string
      enum: [draft, finalized]
    Payroll:
      type: object
      required:
        [
          id,
          employeeId,
          month,
          year,
          baseSalary,
          accruedSalary,
          overtimeAmount,
          bonusesAmount,
          deductionsAmount,
          loanDeduction,
          absenceDeduction,
          finalAmount,
          status,
        ]
      properties:
        id: { type: string }
        employeeId: { type: string }
        employeeName: { type: ["string", "null"] }
        branchId: { type: ["string", "null"] }
        branchName: { type: ["string", "null"] }
        month: { type: integer }
        year: { type: integer }
        baseSalary: { type: number }
        accruedSalary: { type: number }
        overtimeAmount: { type: number }
        bonusesAmount: { type: number }
        deductionsAmount: { type: number }
        loanDeduction: { type: number }
        absenceDeduction: { type: number }
        finalAmount: { type: number }
        status:
          $ref: "#/components/schemas/PayrollStatus"
        generatedAt: { type: ["string", "null"] }
    PayrollGenerateInput:
      type: object
      required: [employeeId, month, year]
      properties:
        employeeId: { type: string }
        month: { type: integer, minimum: 1, maximum: 12 }
        year: { type: integer }
    CurrentPayroll:
      type: object
      required:
        [
          employeeId,
          month,
          year,
          baseSalary,
          daysInMonth,
          daysElapsed,
          accruedSalary,
          accruedPercent,
          overtimeAmount,
          bonusesAmount,
          deductionsAmount,
          loanDeduction,
          absenceDeduction,
          estimatedFinalAmount,
        ]
      properties:
        employeeId: { type: string }
        month: { type: integer }
        year: { type: integer }
        baseSalary: { type: number }
        daysInMonth: { type: integer }
        daysElapsed: { type: integer }
        accruedSalary: { type: number }
        accruedPercent: { type: number }
        overtimeAmount: { type: number }
        bonusesAmount: { type: number }
        deductionsAmount: { type: number }
        loanDeduction: { type: number }
        absenceDeduction: { type: number }
        estimatedFinalAmount: { type: number }
...